Move pico under raspberry pi folder
