+++ src/board.py
@@ -45,7 +45,7 @@ elif board_id == ap_board.PYBOARD:
     from adafruit_blinka.board.pyboard import *
 
 elif board_id == ap_board.RASPBERRY_PI_PICO:
-    from adafruit_blinka.board.raspberry_pi_pico import *
+    from adafruit_blinka.board.raspberrypi.pico import *
 
 elif detector.board.any_raspberry_pi_40_pin:
     from adafruit_blinka.board.raspberrypi.raspi_40pin import *
